Made it easy to add multiple chatrooms to connect to.
Added ChatRoom struct in rooms.rs that stores information about its own
chat room such as id and name of the chatroom, maximum amount of
clients, connected client ids and names as well as chat message history.

When a socket is initially opened, before a "join" message is sent, the
client is sent information about what rooms are available and what their
names are. This should be used for the client to select what room they
want to join. The implementation also allows for users to create a new
chat room if they provide a name/id of a chatroom that doesn't exist.